+++ packages/dash-core-components/src/components/DatePickerRange.react.js
@@ -145,9 +145,10 @@ export default class DatePickerRange extends Component {
                 initialVisibleMonth={() => {
                     if (initial_visible_month) {
                         return initial_visible_month;
-                    }
-                    if (focusedInput === 'endDate') {
+                    } else if (end_date && focusedInput === 'endDate') {
                         return end_date;
+                    } else if (start_date && focusedInput === 'startDate') {
+                        return start_date;
                     }
                     return start_date;
                 }}