+++ packages/dash-core-components/src/components/DatePickerSingle.react.js
@@ -110,7 +110,9 @@ export default class DatePickerSingle extends Component {
                 onDateChange={this.onDateChange}
                 focused={focused}
                 onFocusChange={({focused}) => this.setState({focused})}
-                initialVisibleMonth={() => date || initial_visible_month}
+                initialVisibleMonth={() =>
+                    date || initial_visible_month || moment()
+                }
                 isOutsideRange={this.isOutsideRange}
                 numberOfMonths={number_of_months_shown}
                 withPortal={with_portal && verticalFlag}